+++ packages/verified-fetch/src/plugins/plugin-handle-dag-cbor.ts
@@ -78,11 +78,8 @@ export class DagCborPlugin extends BasePlugin {
     const responseContentType = accept ?? (body instanceof Uint8Array ? 'application/octet-stream' : 'application/json')
     const response = okRangeResponse(resource, context.byteRangeContext.getBody(responseContentType), { byteRangeContext: context.byteRangeContext, log: this.log })
 
-    // if (context.byteRangeContext.isMultiRangeRequest) {
     response.headers.set('content-type', context.byteRangeContext.getContentType() ?? responseContentType)
-    // } else {
-    // response.headers.set('content-type', responseContentType)
-    // }
+
     this.log.trace('setting content type to "%s"', context.byteRangeContext.getContentType() ?? responseContentType)
     setIpfsRoots(response, ipfsRoots)