+++ packages/verified-fetch/test/content-type-parser.spec.ts
@@ -83,7 +83,7 @@ describe('content-type-parser', () => {
     const fs = unixfs(helia)
 
     let barDir = await fs.addDirectory({ path: './bar' })
-    const aFileHtml = await fs.addFile({ path: './bar/a-file.html', content: uint8ArrayFromString('<html><body>Hello world</body></html>') })
+    const aFileHtml = await fs.addBytes(uint8ArrayFromString('<html><body>Hello world</body></html>'))
     barDir = await fs.cp(aFileHtml, barDir, 'a-file.html')
     let fooDir = await fs.addDirectory({ path: './foo' })
     fooDir = await fs.cp(barDir, fooDir, 'bar')